+++ constants/headers.go
@@ -1,0 +1,6 @@
+package constants
+
+const (
+	UserAgentHeader   = "User-Agent"
+	ContentTypeHeader = "Content-Type"
+)