+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.4] - 2025-05-07
+
+### Added
+- Unit tests for `buildURLWithParams` function in requests package
+
+### Fixed
+- Query parameter handling in URLs
+- Custom User Agent in headers
+
 ## [1.3.3] - 2025-05-02
 feat: Added new API endpoints
 - Added Payout API: